#82: cleaning up component, refactoring gamelogs, almost done, just one pesky error message left to deal with
